Fix block break listener if block broken isn't a jukebox

+++ src/main/java/io/github/slazurin/sljukeboxloop/listeners/JukeboxListeners.java
@@ -48,6 +48,10 @@ public class JukeboxListeners implements Listener {
     
     @EventHandler
     public void onJukeboxBreak(BlockBreakEvent event) {
+        if (!event.getBlock().getType().equals(Material.JUKEBOX)) {
+            return;
+        }
+        
         Jukebox j = (Jukebox) event.getBlock().getState();
 
         this.plugin.getApi().delete(j);
